app/requirements.txt
fastapi
uvicorn
motor
numba
strawberry-graphql[fastapi]
temporalio
pytest
pytest-asyncio
jinja2
pyyaml